requirements.txt
pyzbar
pymongo
subprocess

gpiod
Jetson.GPIO==2.1.9
pymongo==4.12.0
Adafruit-Blinka==8.56.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-dht==4.0.7
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11

python-dotenv
asyncio
livekit==0.22.0
livekit-agents==0.12.18
livekit-api==0.8.2
livekit-plugins-google==0.11.2
livekit-protocol==0.9.2
google-genai==1.3.0
google-ai-generativelanguage==0.6.17
traitlets
numpy<2
